chore: move some trace utilities from sw to isomorphic
