+++ packages/playwright-core/src/utils/isomorphic/trace/DEPS.list
@@ -1,0 +1,3 @@
+[*]
+./**
+../**
