+++ packages/playwright-core/src/utils/isomorphic/trace/traceModel.ts
@@ -19,7 +19,7 @@ import { parseClientSideCallMetadata } from '@isomorphic/traceUtils';
 import { SnapshotStorage } from './snapshotStorage';
 import { TraceModernizer } from './traceModernizer';
 
-import type { ContextEntry } from '../types/entries';
+import type { ContextEntry } from './entries';
 
 export interface TraceModelBackend {
   entryNames(): Promise<string[]>;
