+++ packages/playwright-core/src/utils/isomorphic/trace/versions/traceV4.ts
