+++ packages/playwright-core/src/utils/isomorphic/trace/versions/traceV5.ts
