+++ packages/playwright-core/src/utils/isomorphic/trace/versions/traceV6.ts
