+++ packages/playwright-core/src/utils/isomorphic/trace/versions/traceV7.ts
