+++ packages/playwright-core/src/utils/isomorphic/trace/versions/traceV8.ts
