+++ packages/trace/src/trace.ts
@@ -15,7 +15,7 @@
  */
 
 import type { FrameSnapshot, ResourceSnapshot } from './snapshot';
-import type { Language } from '../../playwright-core/src/utils/isomorphic/locatorGenerators';
+import type { Language } from '@isomorphic/locatorGenerators';
 import type { Point, SerializedError, StackFrame } from '@protocol/channels';
 
 export type Size = { width: number, height: number };
